Frameworks: add ability to send a custom intent after keyguard is finished
We have and will have some features which need to start intents after the user
has sucessfully entered the credentials. Add for the system public available method
to start an avtivity after the keyguard.

Add as well support on slim Action for it so that our features actually can use this.

Change-Id: If9b6cbb39f1dd0b3fdad29f3f7b331133d32e89e