Themes: Fix applying theme for notifications
Notifications should be inflated using the current theme applied
to SystemUI.  This was not happening because the Context created
when applying a RemoteView was returning a cahced context rather
than one with the supplied theme applied.

This patch also moves mCurrentTheme into BaseStatusBar so it can
be used when inflating the remote views for notifications.

Change-Id: I577ae30fc02028c7e497cd774f79029f8a045d1f
TICKET: CYNGNOS-1608

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/BaseStatusBar.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java