Hack to remove black bar on custom lockscreen wallpaper
The black bar that we've seen (on Screw'd at least) is the navbar when WallpaperCropper crops the image. It catches the navbar along with the custom image, so you'd see a black bar where the navbar would normally be. That's no bueno.

This puts the view in an "immersive" mode, so the navbar isn't visible when the image is cropped. No navbar when cropping = no black bar on the lockscreen wallpaper.

Used this as a guide on forcing "immersive" view:
https://github.com/ScrewdAOSP/frameworks_base/commit/99752753d55895bf03d9e2572105ed47dd5ba995

So credits to @Mazda-- for unknowingly helping me.

Change-Id: I5391984d9ba2893d2fcd8fd6c9782da1d78e7ec9
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>